// lib/api/model/courseOfferingDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface CourseOfferingDto {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  id: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  courseId: number | string;
  courseCode: string;
  courseTitle: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  termId: number | string;
  termCode: string;
  /** @nullable */
  note: string | null;
  createdAt: string;
}